Location permissions code cleanup
