+++ composeApp/src/wasmJsMain/kotlin/org/noiseplanet/noisecapture/interop/Audio.kt
@@ -1,9 +1,13 @@
-package org.noiseplanet.noisecapture.audio
+package org.noiseplanet.noisecapture.interop
 
 import org.khronos.webgl.Float32Array
 import org.w3c.dom.mediacapture.MediaStream
 import kotlin.js.Promise
 
+/**
+ * AudioContext Kotlin interop.
+ * [MDN Reference](https://developer.mozilla.org/en-US/docs/Web/API/AudioContext)
+ */
 external class AudioContext {
 
     val destination: AudioDestinationNode
@@ -17,14 +21,19 @@ external class AudioContext {
     ): ScriptProcessorNode
 }
 
+/**
+ * AudioDestinationNode Kotlin interop.
+ * [MDN Reference](https://developer.mozilla.org/docs/Web/API/AudioDestinationNode)
+ */
 external class AudioDestinationNode : AudioNode {
 
-    /**
-     * [MDN Reference](https://developer.mozilla.org/docs/Web/API/AudioDestinationNode/maxChannelCount)
-     */
     val maxChannelCount: Int
 }
 
+/**
+ * AudioBuffer Kotlin interop.
+ * [MDN Reference](https://developer.mozilla.org/en-US/docs/Web/API/AudioBuffer)
+ */
 external class AudioBuffer {
 
     val sampleRate: Float
@@ -35,6 +44,10 @@ external class AudioBuffer {
     fun getChannelData(channel: Int): Float32Array
 }
 
+/**
+ * AudioNode Kotlin interop.
+ * [MDN Reference](https://developer.mozilla.org/en-US/docs/Web/API/AudioNode)
+ */
 open external class AudioNode {
 
     fun connect(
@@ -46,11 +59,19 @@ open external class AudioNode {
     fun disconnect()
 }
 
+/**
+ * AudioProcessingEvent Kotlin interop.
+ * [MDN Reference](https://developer.mozilla.org/en-US/docs/Web/API/AudioProcessingEvent)
+ */
 external class AudioProcessingEvent {
 
     val inputBuffer: AudioBuffer
 }
 
+/**
+ * ScriptProcessorNode Kotlin interop.
+ * [MDN Reference](https://developer.mozilla.org/en-US/docs/Web/API/ScriptProcessorNode)
+ */
 external class ScriptProcessorNode : AudioNode {
 
     var onaudioprocess: (AudioProcessingEvent) -> Unit
